reworks hieracy of dto
