+++ openvair/modules/template/domain/model.py
@@ -12,12 +12,12 @@ import abc
 from typing import Dict, ClassVar, cast
 
 from openvair.modules.template.domain.base import BaseTemplate
-from openvair.modules.template.adapters.dto.domain import (
-    DomainTemplateManagerDTO,
-)
 
 # from openvair.modules.template.adapters.dto.templates import TemplateDomain
 from openvair.modules.template.domain.disk_templates.qcow2 import Qcow2Template
+from openvair.modules.template.adapters.dto.internal.models import (
+    DomainTemplateManagerDTO,
+)
 
 
 class AbstractTemplateFactory(metaclass=abc.ABCMeta):
